[vk] Leave comment about the keys for the native image
Currently it is very simple: use the vector's memory address, but
this could lead false assumption as the vector won't change if it
can accomodate the new buffer. Also it would be handy if we add a
API function for updating the image with regions instead of over-
writing the whole image.

Signed-off-by: Kristof Kosztyo <kkristof@inf.u-szeged.hu>